requirements.txt
# Core ML and NLP libraries
tensorflow>=2.10.0
tensorflow-hub>=0.12.0
sentence-transformers>=2.2.0
transformers>=4.20.0
torch>=1.12.0

# Data processing
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
scipy>=1.9.0

# NLP preprocessing
spacy>=3.4.0
nltk>=3.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Dimensionality reduction
umap-learn>=0.5.3

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.64.0
joblib>=1.1.0

# Web interface
streamlit>=1.10.0
fastapi>=0.78.0
uvicorn>=0.18.0

# Testing and development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
pre-commit>=2.20.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
Pillow>=9.0.0
requests>=2.28.0
python-multipart>=0.0.5
